requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
backcall==0.1.0           # via ipython
decorator==4.3.0          # via ipython, traitlets
django-choices==1.6.0
django-cors-headers==2.2.0
django-extensions==2.0.7
django-filter==1.1.0
django==2.0.6
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0
jedi==0.12.0              # via ipython
model-mommy==1.6.0
more-itertools==4.2.0     # via pytest
parso==0.2.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.1.0
pluggy==0.6.0             # via pytest
prompt-toolkit==1.0.15    # via ipython
psycopg2-binary==2.7.4
ptyprocess==0.5.2         # via pexpect
py==1.5.3                 # via pytest
pygments==2.2.0           # via ipython
pyjwt==1.6.4              # via djangorestframework-jwt
pytest-django==3.2.1
pytest==3.6.1             # via pytest-django
pytz==2018.4              # via django
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-extensions, model-mommy, more-itertools, prompt-toolkit, pytest, traitlets
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
